Move request table handling to separate component
